fix(approval): refresh status bar and retire settle duplicates (#9597)
Refresh the extension status-bar policy line after settings/workspace re-seed,
read Tools settings through readPersistedTexraApprovalPolicy, and consolidate
CLI settle helpers into settleApprovals so headless and TUI no longer mirror
decide* glue.
