+++ packages/extension/src/frontend/vscode/texraConfig.ts
@@ -15,6 +15,7 @@ import {
 import { isFileNotFoundError } from '@common/errors';
 
 // Local imports - logging
+import { refreshApprovalPolicyTooltip } from '@frontend/statusBar/approvalPolicyTooltipRefresh';
 import * as logger from '@logger/logUtils';
 
 // Local imports - platform
@@ -150,6 +151,7 @@ export class ExtensionTexraConfig extends JsonConfigProvider {
                 this.get(key, fallback),
               ),
             );
+            refreshApprovalPolicyTooltip();
           },
           afterStorageRollback: rollbackConfig,
           afterStorageFinalize: () => {
